refactor, some styling and making useForm generic to any form
